Move more layout components to components/layout/
